pyproject.toml
[tool.poetry]
name = "notebooks"
version = "0.1.0"
description = ""
authors = ["Zach Morrissey <zachmor05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
sqlalchemy = "^2.0.23"
numpy = "^1.26.2"
psycopg2-binary = "^2.9.9"
seaborn = "^0.13.0"
requests = "^2.31.0"
scipy = "^1.11.4"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.12.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
format = "black ."
